pyproject.toml
[project]
name = "code-agent"
version = "0.1.0"
description = "A Claude Code-like CLI tool for learning purposes"
requires-python = ">=3.12"
dependencies = [
    "textual>=1.0.0",
    "google-genai>=1.0.0",
]

[project.scripts]
code-agent = "code_agent.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/code_agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "textual-dev>=1.0.0",
]
